component separation(CenterProfile)
add TabsSection for component separation
